Add option to control if remote replicas should be used
Previously the driver would always failover to use replicas
in the remote datacenters if there was no replicas available
in local_dc when using DC/RackAwareRoundRobinPolicy. There was
no control over it. This is not how the other driver handle it.

This commit adds a field to dc/rackAwareRR to control if dc
failover is permitted.