Add logic to retry fetching trace info
Previously in gocql there was no mechanism to retry the
session detail fetch, and there is a need for it because
tracing details are stored asynchronously. Other drivers
have it e.g., python-driver and scylla-rust-driver.

This commit adds new logic to be able to configure number
of fetch attempts and intervals between these attempts.