Regenerate testdata/pki and include script for regenerating
The existing certificates in testdata/pki expire on September 16 2024.

This commit includes a 'generate_certs.sh' script for regenerating
private keys and certificates as needed.

As I couldn't find the original steps used to generate these, it's
possible these certificates are generated differently, but they are
done in a nominal way.

One slight derivation with the original certificates is that I have
opted to use PKCS12 format instead of the propertiary java JKS format
for the .truststore and .keystore file.  The cassandra and gocql
certificates also embed a spiffe in the SAN so they can eventually
be used for mTLS authentication testing.

patch by Andy Tolbert; reviewed by Bret McGuire for CASSANDRA-19862