Add new API for non blocking Session creation
For some use-cases it is helpful to be able to start the
initialization of the Session, but not neccesarly wait for
it to finish. To accomodate this need this commit adds new
non blocking API.

If the user tries to query Session that isn't initialized
yet, the new ErrSessionNotReady will be returned.