Switch away from ConnectionTimeout after connection init is done
`ConnectionTimeout` should be applied to any operation on connection
until connection is being initialized.
It goes for both regular and control connection.
When initialization is done we need to switch timeout and writeTimeout
to a `ClusterConfig.Timeout` and `ClusterConfig.WriteTimeout`.

This is all needed to make case when user target low timeout on query
work well.
Without it low timeout will result in inability to create session or
shard connection.